% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSimFunc.R
\name{SimDataWeibFrail}
\alias{SimDataWeibFrail}
\title{Functions simulating data under frailty Weibull illness-death model}
\usage{
SimDataWeibFrail(
  n.sample,
  params,
  no.protected = T,
  no.large = T,
  cens.exp.rate = 0.1,
  round.times = T,
  cens.admin = 100,
  X = NULL
)
}
\description{
Acceptance-rejection sampling is used in case the order preservation assumption is imposed.
}
\author{
Daniel Nevo
}
